Move construct_lambda_schedule from examples into free_energy module
