+++ examples/hif2a/fit_to_multiple_rbfes.py
@@ -11,7 +11,7 @@ from ff.handlers.deserialize import deserialize_handlers
 from ff.handlers import nonbonded
 
 # free energy classes
-from fe.free_energy import RelativeFreeEnergy
+from fe.free_energy import RelativeFreeEnergy, construct_lambda_schedule
 
 # MD initialization
 from md import builders
@@ -74,30 +74,6 @@ forcefield = Forcefield(ff_handlers)
 
 def wrap_method(args, fxn):
     return fxn(*args)
-
-
-def construct_lambda_schedule(num_windows):
-    """manually optimized by YTZ
-
-    TODO: should move this into a common module, probably in free_energy
-    """
-
-    A = int(.35 * num_windows)
-    B = int(.30 * num_windows)
-    C = num_windows - A - B
-
-    # Empirically, we see the largest variance in std <du/dl> near the endpoints in the nonbonded
-    # terms. Bonded terms are roughly linear. So we add more lambda windows at the endpoint to
-    # help improve convergence.
-    lambda_schedule = np.concatenate([
-        np.linspace(0.0, 0.25, A, endpoint=False),
-        np.linspace(0.25, 0.75, B, endpoint=False),
-        np.linspace(0.75, 1.0, C, endpoint=True)
-    ])
-
-    assert len(lambda_schedule) == num_windows
-
-    return lambda_schedule
 
 
 def type_check_handlers(handlers):
