+++ fe/free_energy.py
@@ -465,3 +465,27 @@ class RelativeFreeEnergy(BaseFreeEnergy):
                     grads_and_handles.append((du_dp[0], type(handles)))
 
         return bonded_du_dl, nonbonded_du_dl, grads_and_handles
+
+
+def construct_lambda_schedule(num_windows):
+    """manually optimized by YTZ
+
+    TODO: should move this into a common module, probably in free_energy
+    """
+
+    A = int(.35 * num_windows)
+    B = int(.30 * num_windows)
+    C = num_windows - A - B
+
+    # Empirically, we see the largest variance in std <du/dl> near the endpoints in the nonbonded
+    # terms. Bonded terms are roughly linear. So we add more lambda windows at the endpoint to
+    # help improve convergence.
+    lambda_schedule = np.concatenate([
+        np.linspace(0.0, 0.25, A, endpoint=False),
+        np.linspace(0.25, 0.75, B, endpoint=False),
+        np.linspace(0.75, 1.0, C, endpoint=True)
+    ])
+
+    assert len(lambda_schedule) == num_windows
+
+    return lambda_schedule
